code/backend/requirements.txt
fastapi>=0.100.0,<0.110.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
pydantic>=2.12.5,<3.0.0
pydantic-settings>=2.13.0
PyJWT>=2.8.0
bcrypt>=4.0,<6
# passlib[bcrypt]==1.7.4  -- kept as reference; not called in any code path (bcrypt used directly)
python-multipart==0.0.6
aiosqlite==0.18.0
sqlalchemy[asyncio]==2.0.20
slowapi>=0.1.9
bleach>=6.0.0
